requirements.txt
qdrant-client
fastapi
uvicorn
typing-extensions
python-dotenv
pydantic
langchain
bs4
selenium>=4.0.0
pdfplumber
openai
PyPDF2
python-multipart
-e .